Module fancy_garbling::circuit

DSL for creating circuits compatible with fancy-garbling in the old-fashioned way, where you create a circuit for a computation then garble it.

Structs

ArithmeticCircuit

Static representation of arithmetic computation supported by fancy garbling.

BinaryCircuit

Static representation of binary computation supported by fancy garbling.

CircuitBuilder

CircuitBuilder is used to build circuits.

CircuitRef

The index and modulus of a gate in a circuit.

Enums

ArithmeticGate

Arithmetic computation supported by fancy garbling.

BinaryGate

Binary computation supported by fancy garbling.

Traits

CircuitInfo

Trait to display circuit evaluation costs

CircuitType

Trait representing circuits that can be built by CircuitBuilder

EvaluableCircuit

A Circuit that can be evaluated by a given Fancy object

GateType

Trait representing circuit gates that can be used in CircuitType

Functions

eval_plain

Evaluate the circuit in plaintext.