Struct fancy_garbling::circuit::Circuit

pub struct Circuit { /* fields omitted */ }

Static representation of the type of computation supported by fancy garbling.

Methods

impl Circuit

pub fn new(ngates: Option<usize>) -> Circuit

Make a new Circuit object.

pub fn eval<F: Fancy>(
&self,
f: &mut F,
garbler_inputs: &[F::Item],
evaluator_inputs: &[F::Item]
) -> Result<Option<Vec<u16>>, F::Error>

Evaluate the circuit using fancy object f.

pub fn eval_plain(
&self,
garbler_inputs: &[u16],
evaluator_inputs: &[u16]
) -> Result<Vec<u16>, DummyError>

Evaluate the circuit in plaintext.

pub fn print_info(&self) -> Result<(), DummyError>

Print circuit info.

pub fn num_garbler_inputs(&self) -> usize

Return the number of garbler inputs.

pub fn num_evaluator_inputs(&self) -> usize

Return the number of evaluator inputs.

pub fn noutputs(&self) -> usize

Return the number of outputs.

pub fn modulus(&self, i: usize) -> u16

Return the modulus of the gate indexed by i.

pub fn garbler_input_mod(&self, i: usize) -> u16

Return the modulus of the garbler input indexed by i.

pub fn evaluator_input_mod(&self, i: usize) -> u16

Return the modulus of the evaluator input indexed by i.

impl Circuit

pub fn parse(filename: &str) -> Result<Self, Error>

Generates a new Circuit from file filename. The file must follow the format given here: https://homes.esat.kuleuven.be/~nsmart/MPC/, otherwise a CircuitParserError is returned.