Trait fancy_garbling::circuit::CircuitType

pub trait CircuitType: Clone {
    type Gate: GateType;

    fn increment_nonfree_gates(&mut self);
    fn new(ngates: Option<usize>) -> Self;
    fn get_output_refs(&self) -> &[CircuitRef];
    fn get_garbler_input_refs(&self) -> &[CircuitRef];
    fn get_evaluator_input_refs(&self) -> &[CircuitRef];
    fn get_num_nonfree_gates(&self) -> usize;
    fn push_gates(&mut self, gate: Self::Gate);
    fn push_const_ref(&mut self, xref: CircuitRef);
    fn push_output_ref(&mut self, xref: CircuitRef);
    fn push_garbler_input_ref(&mut self, xref: CircuitRef);
    fn push_evaluator_input_ref(&mut self, xref: CircuitRef);
    fn push_modulus(&mut self, modulus: u16);
    fn garbler_input_mod(&self, i: usize) -> u16;
    fn evaluator_input_mod(&self, i: usize) -> u16;

    fn num_garbler_inputs(&self) -> usize { ... }
    fn num_evaluator_inputs(&self) -> usize { ... }
    fn noutputs(&self) -> usize { ... }

}

Trait representing circuits that can be built by CircuitBuilder

Required Associated Types

type Gate: GateType

Gates that the circuit is composed of

Required Methods

fn increment_nonfree_gates(&mut self)

Increase number of nonfree gates

fn new(ngates: Option<usize>) -> Self

Make a new Circuit object.

fn get_output_refs(&self) -> &[CircuitRef]

Get all output refs

fn get_garbler_input_refs(&self) -> &[CircuitRef]

Get all garbler input refs

fn get_evaluator_input_refs(&self) -> &[CircuitRef]

Get all evaluator input refs

fn get_num_nonfree_gates(&self) -> usize

Get number of nonfree gates

fn push_gates(&mut self, gate: Self::Gate)

Add a gate

fn push_const_ref(&mut self, xref: CircuitRef)

Add a constant ref

fn push_output_ref(&mut self, xref: CircuitRef)

Add an output ref

fn push_garbler_input_ref(&mut self, xref: CircuitRef)

Add a garbler input ref

fn push_evaluator_input_ref(&mut self, xref: CircuitRef)

Add an evaluator input ref

fn push_modulus(&mut self, modulus: u16)

Add wire moulus

fn garbler_input_mod(&self, i: usize) -> u16

Return the modulus of the garbler input indexed by i.

fn evaluator_input_mod(&self, i: usize) -> u16

Return the modulus of the evaluator input indexed by i.

Provided Methods

fn num_garbler_inputs(&self) -> usize

Return the number of garbler inputs.

fn num_evaluator_inputs(&self) -> usize

Return the number of evaluator inputs.

fn noutputs(&self) -> usize

Return the number of outputs.