Trait fancy_garbling::circuit::EvaluableCircuit

pub trait EvaluableCircuit<F: Fancy>: CircuitType {
    fn eval(
        &self,
        f: &mut F,
        garbler_inputs: &[F::Item],
        evaluator_inputs: &[F::Item]
    ) -> Result<Option<Vec<u16>>, F::Error>;
}

A Circuit that can be evaluated by a given Fancy object

Supertrait ensures that circuit can be built by CircuitBuilder

Required Methods

fn eval(
&self,
f: &mut F,
garbler_inputs: &[F::Item],
evaluator_inputs: &[F::Item]
) -> Result<Option<Vec<u16>>, F::Error>

Function to evaluate the circuit