Module fancy_garbling::classic

Provides objects and functions for statically garbling and evaluating a circuit without streaming.

Structs

Encoder
Encode inputs statically.

GarbledCircuit
Static evaluator for a circuit, created by the garble function.

GarbledWriter
Implementation of the Write trait for use by Garbler.

Functions

garble
Garble a circuit without streaming.