Enum fancy_garbling::Wire

pub enum Wire {
    Mod2 {
        val: u128,
    },
    ModN {
        q: u16,
        ds: Vec<u16>,
    },
}

The essential wirelabel type used by garbled circuits.

It is a list of mod-q digits.

Variants
Mod2
Fields of Mod2

val: u128
ModN
Fields of ModN

q: u16ds: Vec<u16>

Methods

impl Wire

pub fn digits(&self) -> Vec<u16>

Get the digits of the wire.

pub fn from_u128(inp: u128, q: u16) -> Self

Get the wire represented by the u128.

pub fn as_u128(&self) -> u128

Pack the wire to a u128.

pub fn zero(q: u16) -> Self

The zero wire for the modulus q.

pub fn rand_delta<R: Rng>(rng: &mut R, q: u16) -> Self

Get a random wire label for mod q, with the first digit set to 1.

pub fn color(&self) -> u16

Get the color digit of the wire.

pub fn plus(&self, other: &Self) -> Self

Add two wires digit-wise mod q, returning a new wire.

pub fn plus_eq<'a>(&'a mut self, other: &Wire) -> &'a mut Wire

Add another wire into this one, digit-wise mod q.

pub fn plus_mov(self, other: &Wire) -> Wire

Add another wire into this one, consuming it for chained computations.

pub fn cmul(&self, c: u16) -> Self

Multiply each digit by a constant c mod q, returning a new wire.

pub fn cmul_eq<'a>(&'a mut self, c: u16) -> &'a mut Wire

Multiply each digit by a constant c mod q.

pub fn cmul_mov(self, c: u16) -> Wire

Multiply each digit by a constant c mod q, consuming it for chained computations.

pub fn negate(&self) -> Self

Negate all the digits mod q, returning a new wire.

pub fn negate_eq<'a>(&'a mut self) -> &'a mut Wire

Negate all the digits mod q.

pub fn negate_mov(self) -> Wire

Negate all the digits mod q, consuming it for chained computations.

pub fn minus(&self, other: &Wire) -> Wire

Subtract two wires, returning the result.

pub fn minus_eq<'a>(&'a mut self, other: &Wire) -> &'a mut Wire

Subtract a wire from this one.

pub fn minus_mov(self, other: &Wire) -> Wire

Subtract a wire from this one, consuming it for chained computations.

pub fn rand<R: Rng>(rng: &mut R, q: u16) -> Wire

Get a random wire mod q.

pub fn hash(&self, tweak: u128) -> u128

Compute the hash of this wire.

Uses fixed-key AES.

pub fn hashback(&self, tweak: u128, new_mod: u16) -> Wire

Compute the hash of this wire, converting the result back to a wire.

Uses fixed-key AES.

pub fn hash2(&self, other: &Wire, tweak: u128) -> u128

Compute the hash of two wires together.

Uses fixed-key AES.

pub fn hashback2(&self, other: &Wire, tweak: u128, new_modulus: u16) -> Wire

Compute the hash of two wires together, converting the result back to a wire.

Uses fixed-key AES.