Enum fancy_garbling::error::FancyError

pub enum FancyError {
    UnequalModuli,
    InvalidArg(String),
    InvalidArgNum {
        got: usize,
        needed: usize,
    },
    InvalidArgMod {
        got: u16,
        needed: u16,
    },
    ArgNotBinary,
    NoTruthTable,
    InvalidTruthTable,
    UninitializedValue,
}

Errors that may occur when using the Fancy trait. These errors are API-usage errors, such as trying to add two Items with different moduli.

Variants
UnequalModuli

Unequal moduli.

InvalidArg(String)

Invalid argument.

InvalidArgNum

Invalid number of arguments.

Fields of InvalidArgNum

got: usize

Received number of arguments.

needed: usize

Expected number of arguments.

InvalidArgMod

Invalid argument modulus.

Fields of InvalidArgMod

got: u16

Received modulus.

needed: u16

Expected modulus.

ArgNotBinary

Expected binary argument.

NoTruthTable

Truth table expected but none given.

InvalidTruthTable

Projection truth table is invalid.

UninitializedValue

Uninitialized value encountered.