Module fancy_garbling::errors

Errors that may be output by this library.

Enums

CircuitBuilderError
Errors emitted when building a circuit.

CircuitParserError
Errors emitted by the circuit parser.

DummyError
Errors from the dummy fancy object.

EvaluatorError
Errors from the evaluator.

FancyError
Errors that may occur when using the Fancy trait. These errors are API-usage errors, such as trying to add two Items with different moduli.

GarblerError
Errors from the garbler.

TwopacError
Errors produced by twopac.