Trait fancy_garbling::fancy::BundleGadgets

pub trait BundleGadgets: Fancy {
    fn garbler_input_bundle(&mut self, ps: &[u16]) -> Bundle<Self::Item> { ... }

    fn evaluator_input_bundle(&mut self, ps: &[u16]) -> Bundle<Self::Item> { ... }

    fn garbler_input_bundle_crt(&mut self, q: u128) -> Bundle<Self::Item> { ... }

    fn evaluator_input_bundle_crt(&mut self, q: u128) -> Bundle<Self::Item> { ... }

    fn constant_bundle(&mut self, xs: &[u16], ps: &[u16]) -> Bundle<Self::Item> { ... }

    fn constant_bundle_crt(&mut self, x: u128, q: u128) -> Bundle<Self::Item> { ... }

    fn garbler_input_bundles(
        &mut self, 
        ps: &[u16], 
        n: usize
    ) -> Vec<Bundle<Self::Item>> { ... }

    fn evaluator_input_bundles(
        &mut self, 
        ps: &[u16], 
        n: usize
    ) -> Vec<Bundle<Self::Item>> { ... }

    fn garbler_input_bundles_crt(
        &mut self, 
        q: u128, 
        n: usize
    ) -> Vec<Bundle<Self::Item>> { ... }

    fn evaluator_input_bundles_crt(
        &mut self, 
        q: u128, 
        n: usize
    ) -> Vec<Bundle<Self::Item>> { ... }

    fn output_bundle(&mut self, x: &Bundle<Self::Item>) { ... }

    fn add_bundles(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Bundle<Self::Item> { ... }

    fn sub_bundles(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Bundle<Self::Item> { ... }

    fn cmul_bundle(
        &mut self, 
        x: &Bundle<Self::Item>, 
        c: u128
    ) -> Bundle<Self::Item> { ... }

    fn mul_bundles(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Bundle<Self::Item> { ... }

    fn cdiv_bundle(
        &mut self, 
        x: &Bundle<Self::Item>, 
        c: u16
    ) -> Bundle<Self::Item> { ... }

    fn cexp_bundle(
        &mut self, 
        x: &Bundle<Self::Item>, 
        c: u16
    ) -> Bundle<Self::Item> { ... }

    fn rem_bundle(
        &mut self, 
        x: &Bundle<Self::Item>, 
        p: u16
    ) -> Bundle<Self::Item> { ... }

    fn eq_bundles(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Self::Item { ... }

    fn mixed_radix_addition(
        &mut self, 
        xs: &[Bundle<Self::Item>]
    ) -> Bundle<Self::Item> { ... }

    fn fractional_mixed_radix(
        &mut self, 
        bun: &Bundle<Self::Item>, 
        ms: &[u16]
    ) -> Bundle<Self::Item> { ... }

    fn relu(
        &mut self, 
        x: &Bundle<Self::Item>, 
        factors_of_m: &[u16]
    ) -> Bundle<Self::Item> { ... }

    fn exact_relu(&mut self, x: &Bundle<Self::Item>) -> Bundle<Self::Item> { ... }

    fn sign(
        &mut self, 
        x: &Bundle<Self::Item>, 
        factors_of_m: &[u16]
    ) -> Self::Item { ... }

    fn exact_sign(&mut self, x: &Bundle<Self::Item>) -> Self::Item { ... }

    fn sgn(
        &mut self, 
        x: &Bundle<Self::Item>, 
        factors_of_m: &[u16]
    ) -> Bundle<Self::Item> { ... }

    fn exact_sgn(&mut self, x: &Bundle<Self::Item>) -> Bundle<Self::Item> { ... }

    fn exact_lt(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Self::Item { ... }

    fn max(&mut self, xs: &[Bundle<Self::Item>]) -> Bundle<Self::Item> { ... }

    fn binary_addition(
        &mut self, 
        xs: &Bundle<Self::Item>, 
        ys: &Bundle<Self::Item>
    ) -> (Bundle<Self::Item>, Self::Item) { ... }

    fn binary_addition_no_carry(
        &mut self, 
        xs: &Bundle<Self::Item>, 
        ys: &Bundle<Self::Item>
    ) -> Bundle<Self::Item> { ... }

    fn twos_complement(&mut self, xs: &Bundle<Self::Item>) -> Bundle<Self::Item> { ... }

    fn binary_subtraction(
        &mut self, 
        xs: &Bundle<Self::Item>, 
        ys: &Bundle<Self::Item>
    ) -> (Bundle<Self::Item>, Self::Item) { ... }

    fn multiplex_constant_bits(
        &mut self, 
        x: &Self::Item, 
        c1: u128, 
        c2: u128, 
        nbits: usize
    ) -> Bundle<Self::Item> { ... }

    fn shift(&mut self, x: &Bundle<Self::Item>, n: usize) -> Bundle<Self::Item> { ... }

    fn binary_cmul(
        &mut self, 
        x: &Bundle<Self::Item>, 
        c: u128, 
        nbits: usize
    ) -> Bundle<Self::Item> { ... }
}

Extension trait for Fancy providing advanced gadgets based on bundles of wires.

Provided Methods

fn garbler_input_bundle(&mut self, ps: &[u16]) -> Bundle<Self::Item>

Crate an input bundle for the garbler using moduli ps.

fn evaluator_input_bundle(&mut self, ps: &[u16]) -> Bundle<Self::Item>

Crate an input bundle for the evaluator using moduli ps.

fn garbler_input_bundle_crt(&mut self, q: u128) -> Bundle<Self::Item>

Crate an input bundle for the garbler using composite CRT modulus q.

fn evaluator_input_bundle_crt(&mut self, q: u128) -> Bundle<Self::Item>

Crate an input bundle for the evaluator using composite CRT modulus q.

fn constant_bundle(&mut self, xs: &[u16], ps: &[u16]) -> Bundle<Self::Item>

Creates a bundle of constant wires using moduli ps.

fn constant_bundle_crt(&mut self, x: u128, q: u128) -> Bundle<Self::Item>

Creates a bundle of constant wires for the CRT representation of x under composite modulus q.

fn garbler_input_bundles(
&mut self,
ps: &[u16],
n: usize
) -> Vec<Bundle<Self::Item>>

Create n garbler input bundles, using moduli ps.

fn evaluator_input_bundles(
&mut self,
ps: &[u16],
n: usize
) -> Vec<Bundle<Self::Item>>

Create n evaluator input bundles, using moduli ps.

fn garbler_input_bundles_crt(
&mut self,
q: u128,
n: usize
) -> Vec<Bundle<Self::Item>>

Create n garbler input bundles, under composite CRT modulus q.

fn evaluator_input_bundles_crt(
&mut self,
q: u128,
n: usize
) -> Vec<Bundle<Self::Item>>

Create n evaluator input bundles, under composite CRT modulus q.

fn output_bundle(&mut self, x: &Bundle<Self::Item>)

Output the wires that make up a bundle.

fn add_bundles(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Bundle<Self::Item>

Add two wire bundles, residue by residue.

fn sub_bundles(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Bundle<Self::Item>

Subtract two wire bundles, residue by residue.

fn cmul_bundle(&mut self, x: &Bundle<Self::Item>, c: u128) -> Bundle<Self::Item>

Multiplies each wire in x by the corresponding residue of c.

fn mul_bundles(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Bundle<Self::Item>

Multiply x with y.

fn cdiv_bundle(&mut self, x: &Bundle<Self::Item>, c: u16) -> Bundle<Self::Item>

Divide x by the constant c. Somewhat finicky, please test. I believe that it requires that c is coprime with all moduli.

fn cexp_bundle(&mut self, x: &Bundle<Self::Item>, c: u16) -> Bundle<Self::Item>

Exponentiate x by the constant c.

fn rem_bundle(&mut self, x: &Bundle<Self::Item>, p: u16) -> Bundle<Self::Item>

Compute the remainder with respect to modulus p.

fn eq_bundles(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Self::Item

Compute x == y. Returns a wire encoding the result mod 2.

fn mixed_radix_addition(
&mut self,
xs: &[Bundle<Self::Item>]
) -> Bundle<Self::Item>

Mixed radix addition.

fn fractional_mixed_radix(
&mut self,
bun: &Bundle<Self::Item>,
ms: &[u16]
) -> Bundle<Self::Item>

Helper function for advanced gadgets, returns the fractional part of X/M where M=product(ms).

fn relu(
&mut self,
x: &Bundle<Self::Item>,
factors_of_m: &[u16]
) -> Bundle<Self::Item>

Compute max(x,0), using potentially approximate factors of M.

fn exact_relu(&mut self, x: &Bundle<Self::Item>) -> Bundle<Self::Item>

Compute max(x,0).

fn sign(&mut self, x: &Bundle<Self::Item>, factors_of_m: &[u16]) -> Self::Item

Return 0 if x is positive and 1 if x is negative. Potentially approximate depending on factors_of_m.

fn exact_sign(&mut self, x: &Bundle<Self::Item>) -> Self::Item

Return 0 if x is positive and 1 if x is negative.

fn sgn(
&mut self,
x: &Bundle<Self::Item>,
factors_of_m: &[u16]
) -> Bundle<Self::Item>

Return if x >= 0 then 1 else -1, where -1 is interpreted as Q-1. Potentially approximate depending on factors_of_m.

fn exact_sgn(&mut self, x: &Bundle<Self::Item>) -> Bundle<Self::Item>

Return if x >= 0 then 1 else -1, where -1 is interpreted as Q-1.

fn exact_lt(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Self::Item

Returns 1 if x < y. Works on both CRT and binary bundles.

fn max(&mut self, xs: &[Bundle<Self::Item>]) -> Bundle<Self::Item>

Compute the maximum bundle in xs. Works on both CRT and binary bundles.

fn binary_addition(
&mut self,
xs: &Bundle<Self::Item>,
ys: &Bundle<Self::Item>
) -> (Bundle<Self::Item>, Self::Item)

Binary addition. Returns the result and the carry.

fn binary_addition_no_carry(
&mut self,
xs: &Bundle<Self::Item>,
ys: &Bundle<Self::Item>
) -> Bundle<Self::Item>

Binary addition. Avoids creating extra gates for the final carry.

fn twos_complement(&mut self, xs: &Bundle<Self::Item>) -> Bundle<Self::Item>

Compute the twos complement of the input bundle (which must be base 2).

fn binary_subtraction(
&mut self,
xs: &Bundle<Self::Item>,
ys: &Bundle<Self::Item>
) -> (Bundle<Self::Item>, Self::Item)

Subtract two binary bundles. Returns the result and whether it overflowed.

fn multiplex_constant_bits(
&mut self,
x: &Self::Item,
c1: u128,
c2: u128,
nbits: usize
) -> Bundle<Self::Item>

If x=0 return c1 as a bundle of constant bits, else return c2.

fn shift(&mut self, x: &Bundle<Self::Item>, n: usize) -> Bundle<Self::Item>

Shift residues, replacing them with zeros in the modulus of the last residue.

fn binary_cmul(
&mut self,
x: &Bundle<Self::Item>,
c: u128,
nbits: usize
) -> Bundle<Self::Item>

Write the constant in binary and that gives you the shift amounts, Eg.. 7x is 4x+2x+x.