Trait fancy_garbling::fancy::Fancy

pub trait Fancy {
    type Item: Clone + Default + HasModulus;
    fn garbler_input(&self, ix: Option<usize>, q: u16) -> Self::Item;

    fn evaluator_input(&self, ix: Option<usize>, q: u16) -> Self::Item;

    fn constant(&self, ix: Option<usize>, x: u16, q: u16) -> Self::Item;

    fn add(&self, x: &Self::Item, y: &Self::Item) -> Self::Item;

    fn sub(&self, x: &Self::Item, y: &Self::Item) -> Self::Item;

    fn cmul(&self, x: &Self::Item, c: u16) -> Self::Item;

    fn mul(
        &self, 
        ix: Option<usize>, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Self::Item;

    fn proj(
        &self, 
        ix: Option<usize>, 
        x: &Self::Item, 
        q: u16, 
        tt: &[u16]
    ) -> Self::Item;

    fn output(&self, ix: Option<usize>, x: &Self::Item);

    fn begin_sync(&self, start_index: usize, end_index: usize);

    fn finish_index(&self, index: usize);

    fn garbler_inputs(
        &self, 
        ix: Option<usize>, 
        q: u16, 
        n: usize
    ) -> Vec<Self::Item> { ... }

    fn evaluator_inputs(
        &self, 
        ix: Option<usize>, 
        q: u16, 
        n: usize
    ) -> Vec<Self::Item> { ... }

    fn add_many(&self, args: &[Self::Item]) -> Self::Item { ... }

    fn xor(&self, x: &Self::Item, y: &Self::Item) -> Self::Item { ... }

    fn negate(&self, ix: Option<usize>, x: &Self::Item) -> Self::Item { ... }

    fn and(
        &self, 
        ix: Option<usize>, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Self::Item { ... }

    fn or(
        &self, 
        ix: Option<usize>, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Self::Item { ... }

    fn and_many(&self, ix: Option<usize>, args: &[Self::Item]) -> Self::Item { ... }

    fn or_many(&self, ix: Option<usize>, args: &[Self::Item]) -> Self::Item { ... }

    fn mod_change(
        &self, 
        ix: Option<usize>, 
        x: &Self::Item, 
        to_modulus: u16
    ) -> Self::Item { ... }

    fn adder(
        &self, 
        ix: Option<usize>, 
        x: &Self::Item, 
        y: &Self::Item, 
        carry_in: Option<&Self::Item>
    ) -> (Self::Item, Self::Item) { ... }

    fn mux(
        &self, 
        ix: Option<usize>, 
        b: &Self::Item, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Self::Item { ... }

    fn mux_constant_bits(
        &self, 
        ix: Option<usize>, 
        x: &Self::Item, 
        b1: bool, 
        b2: bool
    ) -> Self::Item { ... }

    fn outputs(&self, ix: Option<usize>, xs: &[Self::Item]) { ... }
}

DSL for the basic computations supported by fancy-garbling.

Associated Types

type Item: Clone + Default + HasModulus

The underlying wire datatype created by an object implementing Fancy.

Required Methods

fn garbler_input(&self, ix: Option<usize>, q: u16) -> Self::Item

Create an input for the garbler with modulus q.

fn evaluator_input(&self, ix: Option<usize>, q: u16) -> Self::Item

Create an input for the evaluator with modulus q.

fn constant(&self, ix: Option<usize>, x: u16, q: u16) -> Self::Item

Create a constant x with modulus q.

fn add(&self, x: &Self::Item, y: &Self::Item) -> Self::Item

Add x and y.

fn sub(&self, x: &Self::Item, y: &Self::Item) -> Self::Item

Subtract x and y.

fn cmul(&self, x: &Self::Item, c: u16) -> Self::Item

Multiply x times the constant c.

fn mul(&self, ix: Option<usize>, x: &Self::Item, y: &Self::Item) -> Self::Item

Multiply x and y.

fn proj(
&self,
ix: Option<usize>,
x: &Self::Item,
q: u16,
tt: &[u16]
) -> Self::Item

Project x according to the truth table tt. Resulting wire has modulus q.

fn output(&self, ix: Option<usize>, x: &Self::Item)

Process this wire as output.

fn begin_sync(&self, start_index: usize, end_index: usize)

Start synchronization of internal messages.

fn finish_index(&self, index: usize)

Declare this index to be done.

Provided Methods

fn garbler_inputs(&self, ix: Option<usize>, q: u16, n: usize) -> Vec<Self::Item>

Create n garbler inputs with modulus q.

fn evaluator_inputs(
&self,
ix: Option<usize>,
q: u16,
n: usize
) -> Vec<Self::Item>

Create n evaluator inputs with modulus q.

fn add_many(&self, args: &[Self::Item]) -> Self::Item

Sum up a slice of wires.

fn xor(&self, x: &Self::Item, y: &Self::Item) -> Self::Item

Xor is just addition, with the requirement that x and y are mod 2.

fn negate(&self, ix: Option<usize>, x: &Self::Item) -> Self::Item

Negate by xoring x with 1.

fn and(&self, ix: Option<usize>, x: &Self::Item, y: &Self::Item) -> Self::Item

And is just multiplication, with the requirement that x and y are mod 2.

fn or(&self, ix: Option<usize>, x: &Self::Item, y: &Self::Item) -> Self::Item

Or uses Demorgan's Rule implemented with multiplication and negation.

fn and_many(&self, ix: Option<usize>, args: &[Self::Item]) -> Self::Item

Returns 1 if all wires equal 1.

fn or_many(&self, ix: Option<usize>, args: &[Self::Item]) -> Self::Item

Returns 1 if any wire equals 1.

fn mod_change(
&self,
ix: Option<usize>,
x: &Self::Item,
to_modulus: u16
) -> Self::Item

Change the modulus of x to to_modulus using a projection gate.

fn adder(
&self,
ix: Option<usize>,
x: &Self::Item,
y: &Self::Item,
carry_in: Option<&Self::Item>
) -> (Self::Item, Self::Item)

Binary adder. Returns the result and the carry.

fn mux(
&self,
ix: Option<usize>,
b: &Self::Item,
x: &Self::Item,
y: &Self::Item
) -> Self::Item

If b=0 returns x else y.

fn mux_constant_bits(
&self,
ix: Option<usize>,
x: &Self::Item,
b1: bool,
b2: bool
) -> Self::Item

If x=0 return the constant b1 else return b2. Folds constants if possible.

fn outputs(&self, ix: Option<usize>, xs: &[Self::Item])

Output a slice of wires.