Module fancy_garbling::garble

Structs and functions for creating, streaming, and evaluating garbled circuits.

Structs

Decoder
Decode outputs.

Encoder
Encode inputs statically.

Evaluator
Streaming evaluator using a callback to receive ciphertexts as needed.

GarbledCircuit
Static evaluator for a circuit, created by the garble function.

Garbler
Streams garbled circuit ciphertexts through a callback.

Enums

Message
The outputs that can be emitted during streaming of a garbling.

Functions

bench_garbling
Run benchmark garbling and streaming on the function. Garbling function is evaluated on another thread.

garble
Garble a circuit without streaming.

garble_iter
Create an iterator over the messages produced by fancy garbling.

Type Definitions

GarbledGate
The ciphertext created by a garbled gate.

OutputCiphertext
Ciphertext created by the garbler for output gates.