Struct fancy_garbling::garble::Evaluator

pub struct Evaluator<'a> { /* fields omitted */ }

Streaming evaluator using a callback to recieve ciphertexts as needed.

Evaluates a garbled circuit on the fly, using messages containing ciphertexts and wires.

Methods

impl<'a> Evaluator<'a>

pub fn new(recv_function: &mut dyn FnMut() -> Message) -> Evaluator

Create a new Evaluator.

recv_function enables streaming by producing messages during the Fancy computation, which contain ciphertexts and wirelabels.

pub fn decode_output(self) -> Vec<u16>

Decode the output recieved during the Fancy computation.