Struct fancy_garbling::Evaluator

pub struct Evaluator { /* fields omitted */ }

Streaming evaluator using a callback to receive ciphertexts as needed.

Evaluates a garbled circuit on the fly, using messages containing ciphertexts and wires. Parallelizable.

Methods

impl Evaluator

pub fn new<F>(recv_function: F) -> Evaluator where
F: FnMut() -> (Option<SyncIndex>, Message) + Send + 'static,

Create a new Evaluator.

recv_function enables streaming by producing messages during the Fancy computation, which contain ciphertexts and wirelabels.

pub fn decode_output(&self) -> Vec<u16>

Decode the output received during the Fancy computation.