Struct fancy_garbling::Garbler

pub struct Garbler<C, RNG, Wire> { /* private fields */ }

Streams garbled circuit ciphertexts through a callback.

Implementations

impl<C: AbstractChannel, RNG: CryptoRng + RngCore, Wire: WireLabel> Garbler<C, RNG, Wire>

pub fn new(channel: C, rng: RNG) -> Self

Create a new garbler.

pub fn delta(&mut self, q: u16) -> Wire

Create a delta if it has not been created yet for this modulus, otherwise just return the existing one.

pub fn get_deltas(self) -> HashMap<u16, Wire>

Get the deltas, consuming the Garbler.

This is useful for reusing wires in multiple garbled circuit instances.

pub fn send_wire(&mut self, wire: &Wire) -> Result<(), GarblerError>

Send a wire over the established channel.

pub fn encode_wire(&mut self, val: u16, modulus: u16) -> (Wire, Wire)

Encode a wire, producing the zero wire as well as the encoded value.

pub fn encode_many_wires(
&mut self,
vals: &[u16],
moduli: &[u16]
) -> Result<(Vec<Wire>, Vec<Wire>), GarblerError>

Encode many wires, producing zero wires as well as encoded values.

pub fn crt_encode_wire(
&mut self,
val: u128,
modulus: u128
) -> Result<(CrtBundle<Wire>, CrtBundle<Wire>), GarblerError>

Encode a CrtBundle, producing zero wires as well as encoded values.

pub fn bin_encode_wire(
&mut self,
val: u128,
nbits: usize
) -> Result<(BinaryBundle<Wire>, BinaryBundle<Wire>), GarblerError>

Encode a BinaryBundle, producing zero wires as well as encoded values.