Module fancy_garbling::threepac::malicious

Three-party honest-majority multi-party computation with garbled circuits

Implementation of honest-majority three-party malicious secure computation from "Fast and Secure Three-party Computation: The Garbled Circuit Approach". Roughly speaking, the idea is to protect against a malicious garbler by having two garblers, where the evaluator checks that they both send the same data. As an optimization to save bandwidth, we instead have the garblers take turns either sending the garbled circuit or sending a hash of it. The evaluator secret shares its inputs to the two garblers, thus avoiding the need for oblivious transfer.

Usage

There are a few parameters to select. The protocol requires a random number generator and a UniversalHash. In the example below, we will use AesRng and Poly1305. We also need to provide channels to communicate between the parties, and to select how many bytes are sent by one garbler before switching to the other.

const HASH_CHUNK_SIZE: usize = 0x1000;
let mut ev = Evaluator::<UnixChannel, UnixChannel, AesRng, Poly1305>::new(
    ev_channel_gb_1,
    ev_channel_gb_2,
    AesRng::new(),
    HASH_CHUNK_SIZE,
);

Inputs to the garbled circuit are provided through encode, and inputs from other parties are indicated with receive.

let a = ev.encode(input_a, 2)?;            // Input the modulo 2 value input_a.
let b = ev.receive(PartyId::Garbler1, 2)?; // Garbler 1 inputs a modulo 2 value.
let c = ev.receive(PartyId::Garbler2, 2)?;

Each party will run through all of the gates in the circuit, calling a function from Fancy for each gate. Circuits can also be loaded from a file with Circuit::parse(), and executed with eval(). Next, the result is revealed to all parties. Note that FancyReveal::reveal() is used for this, not Fancy::output() which only reveals the result to the evaluator.

let t = ev.and(&a, &b)?;
let r = ev.and(&t, &c)?;
let result = ev.reveal(&r);

The different parties communicate over AbstractChannels, and we will use Unix domain sockets, with each party running in its own thread of the same machine.

let (ev_channel_gb_1, gb_1_channel_ev) = unix_channel_pair();
let (ev_channel_gb_2, gb_2_channel_ev) = unix_channel_pair();
let (gb_1_channel_gb_2, gb_2_channel_gb_1) = unix_channel_pair();

The complete example follows.

const HASH_CHUNK_SIZE: usize = 0x1000;

let (ev_channel_gb_1, gb_1_channel_ev) = unix_channel_pair();
let (ev_channel_gb_2, gb_2_channel_ev) = unix_channel_pair();
let (mut gb_1_channel_gb_2, mut gb_2_channel_gb_1) = unix_channel_pair();
let handle_ev = std::thread::spawn(move || {
    let mut ev = Evaluator::<UnixChannel, UnixChannel, AesRng, Poly1305>::new(
        ev_channel_gb_1,
        ev_channel_gb_2,
        AesRng::new(),
        HASH_CHUNK_SIZE,
    )?;

    let a = ev.encode(input_a, 2)?;
    let b = ev.receive(PartyId::Garbler1, 2)?;
    let c = ev.receive(PartyId::Garbler2, 2)?;

    let t = ev.and(&a, &b)?;
    let r = ev.and(&t, &c)?;
    ev.reveal(&r)
});

let handle_gb_1 = std::thread::spawn(move || {
    let mut gb = Garbler::<UnixChannel, AesRng, Poly1305>::new(
        PartyId::Garbler1,
        &mut gb_1_channel_gb_2,
        gb_1_channel_ev,
        &mut AesRng::new(),
        HASH_CHUNK_SIZE,
    )?;

    let a = gb.receive(PartyId::Evaluator, 2)?;
    let b = gb.encode(input_b, 2)?;
    let c = gb.receive(PartyId::Garbler2, 2)?;

    let t = gb.and(&a, &b)?;
    let r = gb.and(&t, &c)?;
    gb.reveal(&r)
});

let handle_gb_2 = std::thread::spawn(move || {
    let mut gb = Garbler::<UnixChannel, AesRng, Poly1305>::new(
        PartyId::Garbler2,
        &mut gb_2_channel_gb_1,
        gb_2_channel_ev,
        &mut AesRng::new(),
        HASH_CHUNK_SIZE,
    )?;

    let a = gb.receive(PartyId::Evaluator, 2)?;
    let b = gb.receive(PartyId::Garbler1, 2)?;
    let c = gb.encode(input_c, 2)?;

    let t = gb.and(&a, &b)?;
    let r = gb.and(&t, &c)?;
    gb.reveal(&r)
});

let output_ev = handle_ev.join().unwrap().unwrap();
let output_gb_1 = handle_gb_1.join().unwrap().unwrap();
let output_gb_2 = handle_gb_2.join().unwrap().unwrap();

assert_eq!(input_a & input_b & input_c, output_ev);
assert_eq!(output_ev, output_gb_1);
assert_eq!(output_ev, output_gb_2);

Benchmarks

This module utilizes criterion for benchmarking purposes. It is possible to run this benchmark on your own machine using cargo:

$ cargo bench --bench malicious_3pc

Our benchmark uses unix channels to simulate communication between the parties. We ran our benchmark with 10 samples and 100ms of warm up, evaluating AES, SHA-1, and SHA-256 as the benchmark cases.

Following are the summarized results generated by a machine running on: Intel Core i7-8550U CPU @ 1.80GHz and 16 GB of DDR4 RAM.

AES Benchmark

Evaluation of AES encryption between three parties

Lower Bound	Estimate	Upper Bound

Mean	6.2240 ms	6.2356 ms	6.2525 ms

Std. Dev.	3.6750 us	25.533 us	38.340 us

SHA-1 Benchmark

Evaluation of SHA-1 hash between three parties

Lower Bound	Estimate	Upper Bound

Mean	30.695 ms	30.972 ms	31.304 ms

Std. Dev.	230.88 us	526.50 us	720.63 us

SHA-256 Benchmark

Evaluation of SHA-256 hash between three parties

Lower Bound	Estimate	Upper Bound

Mean	55.387 ms	55.509 ms	55.669 ms

Std. Dev.	45.758 us	244.34 us	344.74 us

Modules

evaluator

garbler

Enums

PartyId
Identify a participant in the garbled circuit protocol. There are two garblers and one evaluator.