Trait fancy_garbling::WireLabel

pub trait WireLabel: Clone + HasModulus {

    fn digits(&self) -> Vec<u16>;
    fn as_block(&self) -> Block;
    fn color(&self) -> u16;
    fn plus_eq<'a>(&'a mut self, other: &Self) -> &'a mut Self;
    fn cmul_eq(&mut self, c: u16) -> &mut Self;
    fn negate_eq(&mut self) -> &mut Self;
    fn from_block(inp: Block, q: u16) -> Self;
    fn zero(q: u16) -> Self;
    fn rand_delta<R: CryptoRng + Rng>(rng: &mut R, q: u16) -> Self;
    fn rand<R: CryptoRng + RngCore>(rng: &mut R, q: u16) -> Self;
    fn hash_to_mod(hash: Block, q: u16) -> Self;

    fn hashback(&self, tweak: Block, q: u16) -> Self { ... }
    fn negate_mov(self) -> Self { ... }
    fn cmul_mov(self, c: u16) -> Self { ... }
    fn cmul(&self, c: u16) -> Self { ... }
    fn plus_mov(self, other: &Self) -> Self { ... }
    fn plus(&self, other: &Self) -> Self { ... }
    fn negate(&self) -> Self { ... }
    fn minus_mov(self, other: &Self) -> Self { ... }
    fn minus(&self, other: &Self) -> Self { ... }
    fn minus_eq<'a>(&'a mut self, other: &Self) -> &'a mut Self { ... }
    fn hash(&self, tweak: Block) -> Block { ... }

}

Trait implementing a wire that can be used for secure computation via garbled circuits

Required Methods

fn digits(&self) -> Vec<u16>

Get the digits of the wire

fn as_block(&self) -> Block

Pack the wire into a Block.

fn color(&self) -> u16

Get the color digit of the wire.

fn plus_eq<'a>(&'a mut self, other: &Self) -> &'a mut Self

Add another wire digit-wise into this one. Assumes that both wires have the same modulus.

fn cmul_eq(&mut self, c: u16) -> &mut Self

Multiply each digit by a constant c mod q.

fn negate_eq(&mut self) -> &mut Self

Negate all the digits mod q.

fn from_block(inp: Block, q: u16) -> Self

Pack the wire into a Block.

fn zero(q: u16) -> Self

The zero wire with modulus q

fn rand_delta<R: CryptoRng + Rng>(rng: &mut R, q: u16) -> Self

Get a random wire label mod q, with the first digit set to 1

fn rand<R: CryptoRng + RngCore>(rng: &mut R, q: u16) -> Self

Get a random wire mod q.

fn hash_to_mod(hash: Block, q: u16) -> Self

Subroutine of hashback that converts the hash block into a valid wire of the given modulus. Also useful when batching hashes ahead of time for later conversion.

Provided Methods

fn hashback(&self, tweak: Block, q: u16) -> Self

Compute the hash of this wire, converting the result back to a wire.

Uses fixed-key AES.

fn negate_mov(self) -> Self

Negate all the digits mod q, consuming it for chained computations.

fn cmul_mov(self, c: u16) -> Self

Multiply each digit by a constant c mod q, consuming it for chained computations.

fn cmul(&self, c: u16) -> Self

Multiply each digit by a constant c mod q, returning a new wire.

fn plus_mov(self, other: &Self) -> Self

Add another wire into this one, consuming it for chained computations.

fn plus(&self, other: &Self) -> Self

Add two wires digit-wise, returning a new wire.

fn negate(&self) -> Self

Negate all the digits mod q, returning a new wire.

fn minus_mov(self, other: &Self) -> Self

Subtract a wire from this one, consuming it for chained computations.

fn minus(&self, other: &Self) -> Self

Subtract two wires, returning the result.

fn minus_eq<'a>(&'a mut self, other: &Self) -> &'a mut Self

Subtract a wire from this one.

fn hash(&self, tweak: Block) -> Block

Compute the hash of this wire.

Uses fixed-key AES.