Crate fancy_garbling

Modules

aes
Wrapper for aesni with constant key for use as a hash function.

circuit
DSL for creating circuits compatible with fancy-garbling in the old-fashioned way, where you create a circuit for a computation then garble it.

dummy
Dummy implementation of Fancy.

fancy
The Fancy trait represents the kinds of computations possible in fancy-garbling.

garble
Structs and functions for creating, streaming, and evaluating garbled circuits.

informer
Informer runs a fancy computation and learns information from it, like how many of what kind of inputs there are.

util
Tools useful for interacting with fancy-garbling.

wire
Low-level operations on wirelabels, the basic building block of garbled circuits.