Enum fancy_garbling::error::FancyError

pub enum FancyError<T> {
    UnequalModuli,
    NotImplemented,
    InvalidArg {
        desc: String,
    },
    InvalidArgNum {
        got: usize,
        needed: usize,
    },
    InvalidArgMod {
        got: u16,
        needed: u16,
    },
    ArgNotBinary,
    NoTruthTable,
    InvalidTruthTable,
    UninitializedValue,
    ClientError(T),
}

Errors that may occur when using the Fancy trait. ClientError wraps errors from underlying fancy object.

With the exception of ClientError, FancyErrors should be API-usage errors, such as trying to add two Items with different moduli.

Variants
UnequalModuliNotImplementedInvalidArg
Fields of InvalidArg

desc: String
InvalidArgNum
Fields of InvalidArgNum

got: usizeneeded: usize
InvalidArgMod
Fields of InvalidArgMod

got: u16needed: u16
ArgNotBinaryNoTruthTableInvalidTruthTableUninitializedValueClientError(T)
Methods

impl<A> FancyError<A>

pub fn map_client_err<B>(
self,
f: impl Fn(A) -> B
) -> FancyError<B>

Map the underlying client error type.

Useful for wrapping errors from fancy objects.