Trait fancy_garbling::fancy::Fancy

pub trait Fancy {
    type Item: Clone + Default + HasModulus;
    fn garbler_input(&mut self, q: u16) -> Self::Item;

    fn evaluator_input(&mut self, q: u16) -> Self::Item;

    fn constant(&mut self, x: u16, q: u16) -> Self::Item;

    fn add(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item;

    fn sub(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item;

    fn cmul(&mut self, x: &Self::Item, c: u16) -> Self::Item;

    fn mul(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item;

    fn proj(&mut self, x: &Self::Item, q: u16, tt: &[u16]) -> Self::Item;

    fn output(&mut self, x: &Self::Item);

    fn garbler_inputs(&mut self, q: u16, n: usize) -> Vec<Self::Item> { ... }

    fn evaluator_inputs(&mut self, q: u16, n: usize) -> Vec<Self::Item> { ... }

    fn add_many(&mut self, args: &[Self::Item]) -> Self::Item { ... }

    fn xor(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item { ... }

    fn negate(&mut self, x: &Self::Item) -> Self::Item { ... }

    fn and(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item { ... }

    fn or(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item { ... }

    fn and_many(&mut self, args: &[Self::Item]) -> Self::Item { ... }

    fn or_many(&mut self, args: &[Self::Item]) -> Self::Item { ... }

    fn mod_change(&mut self, x: &Self::Item, to_modulus: u16) -> Self::Item { ... }

    fn adder(
        &mut self, 
        x: &Self::Item, 
        y: &Self::Item, 
        carry_in: Option<&Self::Item>
    ) -> (Self::Item, Self::Item) { ... }

    fn mux(&mut self, x: &Self::Item, b1: bool, b2: bool) -> Self::Item { ... }

    fn outputs(&mut self, xs: &[Self::Item]) { ... }
}

DSL for the basic computations supported by fancy-garbling.

Associated Types

type Item: Clone + Default + HasModulus

The underlying wire datatype created by an object implementing Fancy.

Required Methods

fn garbler_input(&mut self, q: u16) -> Self::Item

Create an input for the garbler with modulus q.

fn evaluator_input(&mut self, q: u16) -> Self::Item

Create an input for the evaluator with modulus q.

fn constant(&mut self, x: u16, q: u16) -> Self::Item

Create a constant x with modulus q.

fn add(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item

Add x and y.

fn sub(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item

Subtract x and y.

fn cmul(&mut self, x: &Self::Item, c: u16) -> Self::Item

Multiply x times the constant c.

fn mul(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item

Multiply x and y.

fn proj(&mut self, x: &Self::Item, q: u16, tt: &[u16]) -> Self::Item

Project x according to the truth table tt. Resulting wire has modulus q.

fn output(&mut self, x: &Self::Item)

Process this wire as output.

Provided Methods

fn garbler_inputs(&mut self, q: u16, n: usize) -> Vec<Self::Item>

Create n garbler inputs with modulus q.

fn evaluator_inputs(&mut self, q: u16, n: usize) -> Vec<Self::Item>

Create n evaluator inputs with modulus q.

fn add_many(&mut self, args: &[Self::Item]) -> Self::Item

Sum up a slice of wires.

fn xor(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item

Xor is just addition, with the requirement that x and y are mod 2.

fn negate(&mut self, x: &Self::Item) -> Self::Item

Negate by xoring x with 1.

fn and(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item

And is just multiplication, with the requirement that x and y are mod 2.

fn or(&mut self, x: &Self::Item, y: &Self::Item) -> Self::Item

Or uses Demorgan's Rule implemented with multiplication and negation.

fn and_many(&mut self, args: &[Self::Item]) -> Self::Item

Returns 1 if all wires equal 1.

fn or_many(&mut self, args: &[Self::Item]) -> Self::Item

Returns 1 if any wire equals 1.

fn mod_change(&mut self, x: &Self::Item, to_modulus: u16) -> Self::Item

Change the modulus of x to to_modulus using a projection gate.

fn adder(
&mut self,
x: &Self::Item,
y: &Self::Item,
carry_in: Option<&Self::Item>
) -> (Self::Item, Self::Item)

Binary adder. Returns the result and the carry.

fn mux(&mut self, x: &Self::Item, b1: bool, b2: bool) -> Self::Item

If x=0 return the constant b1 else return b2. Folds constants if possible.

fn outputs(&mut self, xs: &[Self::Item])

Output a slice of wires.