Struct fancy_garbling::Evaluator

pub struct Evaluator<R: Read + Debug> { /* fields omitted */ }

Streaming evaluator using a callback to receive ciphertexts as needed.

Evaluates a garbled circuit on the fly, using messages containing ciphertexts and wires. Parallelizable.

Methods

impl<R: Read + Debug> Evaluator<R>

pub fn new(reader: Rc<RefCell<R>>) -> Self

Create a new Evaluator.

callback enables streaming by producing messages during the Fancy computation, which contain ciphertexts and wire-labels.

pub fn decode_output(&self) -> Vec<u16>

Decode the output received during the Fancy computation.