Trait fancy_garbling::BundleGadgets

pub trait BundleGadgets: Fancy {
    fn garbler_input_bundle(
        &mut self, 
        ps: &[u16], 
        opt_xs: Option<Vec<u16>>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn evaluator_input_bundle(
        &mut self, 
        ps: &[u16]
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn garbler_input_bundle_crt(
        &mut self, 
        q: u128, 
        opt_x: Option<u128>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn evaluator_input_bundle_crt(
        &mut self, 
        q: u128
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn garbler_input_bundle_binary(
        &mut self, 
        nbits: usize, 
        opt_x: Option<u128>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn evaluator_input_bundle_binary(
        &mut self, 
        n: usize
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn constant_bundle(
        &mut self, 
        xs: &[u16], 
        ps: &[u16]
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn constant_bundle_crt(
        &mut self, 
        x: u128, 
        q: u128
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn constant_bundle_binary(
        &mut self, 
        val: u128, 
        nbits: usize
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn garbler_input_bundles(
        &mut self, 
        ps: &[u16], 
        n: usize, 
        opt_xs: Option<Vec<Vec<u16>>>
    ) -> Result<Vec<Bundle<Self::Item>>, Self::Error> { ... }

    fn evaluator_input_bundles(
        &mut self, 
        ps: &[u16], 
        n: usize
    ) -> Result<Vec<Bundle<Self::Item>>, Self::Error> { ... }

    fn garbler_input_bundles_crt(
        &mut self, 
        q: u128, 
        n: usize, 
        opt_xs: Option<Vec<u128>>
    ) -> Result<Vec<Bundle<Self::Item>>, Self::Error> { ... }

    fn evaluator_input_bundles_crt(
        &mut self, 
        q: u128, 
        n: usize
    ) -> Result<Vec<Bundle<Self::Item>>, Self::Error> { ... }

    fn output_bundle(
        &mut self, 
        x: &Bundle<Self::Item>
    ) -> Result<(), Self::Error> { ... }

    fn output_bundles(
        &mut self, 
        xs: &[Bundle<Self::Item>]
    ) -> Result<(), Self::Error> { ... }

    fn add_bundles(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn sub_bundles(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn cmul_bundle(
        &mut self, 
        x: &Bundle<Self::Item>, 
        c: u128
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn mul_bundles(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn cexp_bundle(
        &mut self, 
        x: &Bundle<Self::Item>, 
        c: u16
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn rem_bundle(
        &mut self, 
        x: &Bundle<Self::Item>, 
        p: u16
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn eq_bundles(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Result<Self::Item, Self::Error> { ... }

    fn mixed_radix_addition(
        &mut self, 
        xs: &[Bundle<Self::Item>]
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn mixed_radix_addition_msb_only(
        &mut self, 
        xs: &[Bundle<Self::Item>]
    ) -> Result<Self::Item, Self::Error> { ... }

    fn fractional_mixed_radix(
        &mut self, 
        bun: &Bundle<Self::Item>, 
        ms: &[u16]
    ) -> Result<Self::Item, Self::Error> { ... }

    fn relu(
        &mut self, 
        x: &Bundle<Self::Item>, 
        accuracy: &str, 
        output_moduli: Option<&[u16]>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn sign(
        &mut self, 
        x: &Bundle<Self::Item>, 
        accuracy: &str
    ) -> Result<Self::Item, Self::Error> { ... }

    fn sgn(
        &mut self, 
        x: &Bundle<Self::Item>, 
        accuracy: &str, 
        output_moduli: Option<&[u16]>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn lt(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>, 
        accuracy: &str
    ) -> Result<Self::Item, Self::Error> { ... }

    fn geq(
        &mut self, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>, 
        accuracy: &str
    ) -> Result<Self::Item, Self::Error> { ... }

    fn max(
        &mut self, 
        xs: &[Bundle<Self::Item>], 
        accuracy: &str
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn binary_addition(
        &mut self, 
        xs: &Bundle<Self::Item>, 
        ys: &Bundle<Self::Item>
    ) -> Result<(Bundle<Self::Item>, Self::Item), Self::Error> { ... }

    fn binary_addition_no_carry(
        &mut self, 
        xs: &Bundle<Self::Item>, 
        ys: &Bundle<Self::Item>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn binary_multiplication_lower_half(
        &mut self, 
        xs: &Bundle<Self::Item>, 
        ys: &Bundle<Self::Item>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn twos_complement(
        &mut self, 
        xs: &Bundle<Self::Item>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn binary_subtraction(
        &mut self, 
        xs: &Bundle<Self::Item>, 
        ys: &Bundle<Self::Item>
    ) -> Result<(Bundle<Self::Item>, Self::Item), Self::Error> { ... }

    fn multiplex(
        &mut self, 
        b: &Self::Item, 
        x: &Bundle<Self::Item>, 
        y: &Bundle<Self::Item>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn multiplex_constant_bits(
        &mut self, 
        x: &Self::Item, 
        c1: u128, 
        c2: u128, 
        nbits: usize
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn shift(
        &mut self, 
        x: &Bundle<Self::Item>, 
        n: usize
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn binary_cmul(
        &mut self, 
        x: &Bundle<Self::Item>, 
        c: u128, 
        nbits: usize
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }

    fn abs(
        &mut self, 
        x: &Bundle<Self::Item>
    ) -> Result<Bundle<Self::Item>, Self::Error> { ... }
}

Extension trait for Fancy providing advanced gadgets based on bundles of wires.

Provided methods

fn garbler_input_bundle(
&mut self,
ps: &[u16],
opt_xs: Option<Vec<u16>>
) -> Result<Bundle<Self::Item>, Self::Error>

Crate an input bundle for the garbler using moduli ps and optional inputs xs.

fn evaluator_input_bundle(
&mut self,
ps: &[u16]
) -> Result<Bundle<Self::Item>, Self::Error>

Crate an input bundle for the evaluator using moduli ps.

fn garbler_input_bundle_crt(
&mut self,
q: u128,
opt_x: Option<u128>
) -> Result<Bundle<Self::Item>, Self::Error>

Crate an input bundle for the garbler using composite CRT modulus q and optional input x.

fn evaluator_input_bundle_crt(
&mut self,
q: u128
) -> Result<Bundle<Self::Item>, Self::Error>

Crate an input bundle for the evaluator using composite CRT modulus q.

fn garbler_input_bundle_binary(
&mut self,
nbits: usize,
opt_x: Option<u128>
) -> Result<Bundle<Self::Item>, Self::Error>

Create an input bundle for the garbler using nbits base 2 inputs and optional input x.

fn evaluator_input_bundle_binary(
&mut self,
n: usize
) -> Result<Bundle<Self::Item>, Self::Error>

Create an input bundle for the evaluator using n base 2 inputs.

fn constant_bundle(
&mut self,
xs: &[u16],
ps: &[u16]
) -> Result<Bundle<Self::Item>, Self::Error>

Creates a bundle of constant wires using moduli ps.

fn constant_bundle_crt(
&mut self,
x: u128,
q: u128
) -> Result<Bundle<Self::Item>, Self::Error>

Creates a bundle of constant wires for the CRT representation of x under composite modulus q.

fn constant_bundle_binary(
&mut self,
val: u128,
nbits: usize
) -> Result<Bundle<Self::Item>, Self::Error>

Create a constant bundle using base 2 inputs.

fn garbler_input_bundles(
&mut self,
ps: &[u16],
n: usize,
opt_xs: Option<Vec<Vec<u16>>>
) -> Result<Vec<Bundle<Self::Item>>, Self::Error>

Create n garbler input bundles, using moduli ps and optional inputs xs.

fn evaluator_input_bundles(
&mut self,
ps: &[u16],
n: usize
) -> Result<Vec<Bundle<Self::Item>>, Self::Error>

Create n evaluator input bundles, using moduli ps.

fn garbler_input_bundles_crt(
&mut self,
q: u128,
n: usize,
opt_xs: Option<Vec<u128>>
) -> Result<Vec<Bundle<Self::Item>>, Self::Error>

Create n garbler input bundles, under composite CRT modulus q and optional inputs xs.

fn evaluator_input_bundles_crt(
&mut self,
q: u128,
n: usize
) -> Result<Vec<Bundle<Self::Item>>, Self::Error>

Create n evaluator input bundles, under composite CRT modulus q.

fn output_bundle(&mut self, x: &Bundle<Self::Item>) -> Result<(), Self::Error>

Output the wires that make up a bundle.

fn output_bundles(
&mut self,
xs: &[Bundle<Self::Item>]
) -> Result<(), Self::Error>

Output a slice of bundles.

fn add_bundles(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Result<Bundle<Self::Item>, Self::Error>

Add two wire bundles, residue by residue.

fn sub_bundles(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Result<Bundle<Self::Item>, Self::Error>

Subtract two wire bundles, residue by residue.

fn cmul_bundle(
&mut self,
x: &Bundle<Self::Item>,
c: u128
) -> Result<Bundle<Self::Item>, Self::Error>

Multiplies each wire in x by the corresponding residue of c.

fn mul_bundles(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Result<Bundle<Self::Item>, Self::Error>

Multiply x with y.

fn cexp_bundle(
&mut self,
x: &Bundle<Self::Item>,
c: u16
) -> Result<Bundle<Self::Item>, Self::Error>

Exponentiate x by the constant c.

fn rem_bundle(
&mut self,
x: &Bundle<Self::Item>,
p: u16
) -> Result<Bundle<Self::Item>, Self::Error>

Compute the remainder with respect to modulus p.

fn eq_bundles(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Result<Self::Item, Self::Error>

Compute x == y. Returns a wire encoding the result mod 2.

fn mixed_radix_addition(
&mut self,
xs: &[Bundle<Self::Item>]
) -> Result<Bundle<Self::Item>, Self::Error>

Mixed radix addition.

fn mixed_radix_addition_msb_only(
&mut self,
xs: &[Bundle<Self::Item>]
) -> Result<Self::Item, Self::Error>

Mixed radix addition only returning the MSB.

fn fractional_mixed_radix(
&mut self,
bun: &Bundle<Self::Item>,
ms: &[u16]
) -> Result<Self::Item, Self::Error>

Helper function for advanced gadgets, returns the MSB of the fractional part of X/M where M=product(ms).

fn relu(
&mut self,
x: &Bundle<Self::Item>,
accuracy: &str,
output_moduli: Option<&[u16]>
) -> Result<Bundle<Self::Item>, Self::Error>

Compute max(x,0).

Optional output moduli.

fn sign(
&mut self,
x: &Bundle<Self::Item>,
accuracy: &str
) -> Result<Self::Item, Self::Error>

Return 0 if x is positive and 1 if x is negative.

fn sgn(
&mut self,
x: &Bundle<Self::Item>,
accuracy: &str,
output_moduli: Option<&[u16]>
) -> Result<Bundle<Self::Item>, Self::Error>

Return if x >= 0 then 1 else -1, where -1 is interpreted as Q-1.

If provided, will produce a bundle under output_moduli instead of x.moduli()

fn lt(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>,
accuracy: &str
) -> Result<Self::Item, Self::Error>

Returns 1 if x < y. Works on both CRT and binary bundles.

Binary ignores accuracy argument.

fn geq(
&mut self,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>,
accuracy: &str
) -> Result<Self::Item, Self::Error>

Returns 1 if x >= y. Works on both CRT and binary bundles.

fn max(
&mut self,
xs: &[Bundle<Self::Item>],
accuracy: &str
) -> Result<Bundle<Self::Item>, Self::Error>

Compute the maximum bundle in xs.

fn binary_addition(
&mut self,
xs: &Bundle<Self::Item>,
ys: &Bundle<Self::Item>
) -> Result<(Bundle<Self::Item>, Self::Item), Self::Error>

Binary addition. Returns the result and the carry.

fn binary_addition_no_carry(
&mut self,
xs: &Bundle<Self::Item>,
ys: &Bundle<Self::Item>
) -> Result<Bundle<Self::Item>, Self::Error>

Binary addition. Avoids creating extra gates for the final carry.

fn binary_multiplication_lower_half(
&mut self,
xs: &Bundle<Self::Item>,
ys: &Bundle<Self::Item>
) -> Result<Bundle<Self::Item>, Self::Error>

Binary multiplication.

Returns the lower-order half of the output bits, ie a number with the same number of bits as the inputs.

fn twos_complement(
&mut self,
xs: &Bundle<Self::Item>
) -> Result<Bundle<Self::Item>, Self::Error>

Compute the twos complement of the input bundle (which must be base 2).

fn binary_subtraction(
&mut self,
xs: &Bundle<Self::Item>,
ys: &Bundle<Self::Item>
) -> Result<(Bundle<Self::Item>, Self::Item), Self::Error>

Subtract two binary bundles. Returns the result and whether it underflowed.

Due to the way that twos_complement(0) = 0, underflow indicates y != 0 && x >= y.

fn multiplex(
&mut self,
b: &Self::Item,
x: &Bundle<Self::Item>,
y: &Bundle<Self::Item>
) -> Result<Bundle<Self::Item>, Self::Error>

If b=0 then return x, else return y.

fn multiplex_constant_bits(
&mut self,
x: &Self::Item,
c1: u128,
c2: u128,
nbits: usize
) -> Result<Bundle<Self::Item>, Self::Error>

If x=0 return c1 as a bundle of constant bits, else return c2.

fn shift(
&mut self,
x: &Bundle<Self::Item>,
n: usize
) -> Result<Bundle<Self::Item>, Self::Error>

Shift residues, replacing them with zeros in the modulus of the least signifigant residue.

fn binary_cmul(
&mut self,
x: &Bundle<Self::Item>,
c: u128,
nbits: usize
) -> Result<Bundle<Self::Item>, Self::Error>

Write the constant in binary and that gives you the shift amounts, Eg.. 7x is 4x+2x+x.

fn abs(
&mut self,
x: &Bundle<Self::Item>
) -> Result<Bundle<Self::Item>, Self::Error>

Compute the absolute value of a binary bundle.

Loading content...