Trait fancy_garbling::Fancy

pub trait Fancy {
    type Item: Clone + HasModulus;
    type Error: Debug + Display;
    fn garbler_input(
        &self, 
        ix: Option<SyncIndex>, 
        q: u16, 
        opt_x: Option<u16>
    ) -> Result<Self::Item, FancyError<Self::Error>>;

    fn evaluator_input(
        &self, 
        ix: Option<SyncIndex>, 
        q: u16
    ) -> Result<Self::Item, FancyError<Self::Error>>;

    fn constant(
        &self, 
        ix: Option<SyncIndex>, 
        x: u16, 
        q: u16
    ) -> Result<Self::Item, FancyError<Self::Error>>;

    fn add(
        &self, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Result<Self::Item, FancyError<Self::Error>>;

    fn sub(
        &self, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Result<Self::Item, FancyError<Self::Error>>;

    fn cmul(
        &self, 
        x: &Self::Item, 
        c: u16
    ) -> Result<Self::Item, FancyError<Self::Error>>;

    fn mul(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Result<Self::Item, FancyError<Self::Error>>;

    fn proj(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item, 
        q: u16, 
        tt: Option<Vec<u16>>
    ) -> Result<Self::Item, FancyError<Self::Error>>;

    fn output(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item
    ) -> Result<(), FancyError<Self::Error>>;

    fn begin_sync(
        &self, 
        _num_indices: SyncIndex
    ) -> Result<(), FancyError<Self::Error>> { ... }

    fn finish_index(
        &self, 
        _ix: SyncIndex
    ) -> Result<(), FancyError<Self::Error>> { ... }

    fn garbler_inputs(
        &self, 
        ix: Option<SyncIndex>, 
        qs: &[u16], 
        opt_xs: Option<Vec<u16>>
    ) -> Result<Vec<Self::Item>, FancyError<Self::Error>> { ... }

    fn evaluator_inputs(
        &self, 
        ix: Option<SyncIndex>, 
        qs: &[u16]
    ) -> Result<Vec<Self::Item>, FancyError<Self::Error>> { ... }

    fn add_many(
        &self, 
        args: &[Self::Item]
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn xor(
        &self, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn negate(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn and(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn or(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn and_many(
        &self, 
        ix: Option<SyncIndex>, 
        args: &[Self::Item]
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn or_many(
        &self, 
        ix: Option<SyncIndex>, 
        args: &[Self::Item]
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn mod_change(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item, 
        to_modulus: u16
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn adder(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item, 
        y: &Self::Item, 
        carry_in: Option<&Self::Item>
    ) -> Result<(Self::Item, Self::Item), FancyError<Self::Error>> { ... }

    fn mux(
        &self, 
        ix: Option<SyncIndex>, 
        b: &Self::Item, 
        x: &Self::Item, 
        y: &Self::Item
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn mux_constant_bits(
        &self, 
        ix: Option<SyncIndex>, 
        x: &Self::Item, 
        b1: bool, 
        b2: bool
    ) -> Result<Self::Item, FancyError<Self::Error>> { ... }

    fn outputs(
        &self, 
        ix: Option<SyncIndex>, 
        xs: &[Self::Item]
    ) -> Result<(), FancyError<Self::Error>> { ... }
}

DSL for the basic computations supported by fancy-garbling.

Associated Types

type Item: Clone + HasModulus

The underlying wire datatype created by an object implementing Fancy.

type Error: Debug + Display

Errors which may be thrown by the users of Fancy.

Required Methods

fn garbler_input(
&self,
ix: Option<SyncIndex>,
q: u16,
opt_x: Option<u16>
) -> Result<Self::Item, FancyError<Self::Error>>

Create an input for the garbler with modulus q and optional garbler-private value x.

fn evaluator_input(
&self,
ix: Option<SyncIndex>,
q: u16
) -> Result<Self::Item, FancyError<Self::Error>>

Create an input for the evaluator with modulus q.

fn constant(
&self,
ix: Option<SyncIndex>,
x: u16,
q: u16
) -> Result<Self::Item, FancyError<Self::Error>>

Create a constant x with modulus q.

fn add(
&self,
x: &Self::Item,
y: &Self::Item
) -> Result<Self::Item, FancyError<Self::Error>>

Add x and y.

fn sub(
&self,
x: &Self::Item,
y: &Self::Item
) -> Result<Self::Item, FancyError<Self::Error>>

Subtract x and y.

fn cmul(
&self,
x: &Self::Item,
c: u16
) -> Result<Self::Item, FancyError<Self::Error>>

Multiply x times the constant c.

fn mul(
&self,
ix: Option<SyncIndex>,
x: &Self::Item,
y: &Self::Item
) -> Result<Self::Item, FancyError<Self::Error>>

Multiply x and y.

fn proj(
&self,
ix: Option<SyncIndex>,
x: &Self::Item,
q: u16,
tt: Option<Vec<u16>>
) -> Result<Self::Item, FancyError<Self::Error>>

Project x according to the truth table tt. Resulting wire has modulus q.

Optional tt is useful for hiding the gate from evaluator.

fn output(
&self,
ix: Option<SyncIndex>,
x: &Self::Item
) -> Result<(), FancyError<Self::Error>>

Process this wire as output.

Provided Methods

fn begin_sync(
&self,
_num_indices: SyncIndex
) -> Result<(), FancyError<Self::Error>>

Start synchronization of internal messages.

Optional, throws FancyError::NotImplemented if used without an implementation.

fn finish_index(&self, _ix: SyncIndex) -> Result<(), FancyError<Self::Error>>

Declare this index to be done.

Optional, throws FancyError::NotImplemented if used without an implementation.

fn garbler_inputs(
&self,
ix: Option<SyncIndex>,
qs: &[u16],
opt_xs: Option<Vec<u16>>
) -> Result<Vec<Self::Item>, FancyError<Self::Error>>

Create n garbler inputs with the moduli qs and optional inputs xs.

fn evaluator_inputs(
&self,
ix: Option<SyncIndex>,
qs: &[u16]
) -> Result<Vec<Self::Item>, FancyError<Self::Error>>

Create n evaluator inputs with the moduli qs.

fn add_many(
&self,
args: &[Self::Item]
) -> Result<Self::Item, FancyError<Self::Error>>

Sum up a slice of wires.

fn xor(
&self,
x: &Self::Item,
y: &Self::Item
) -> Result<Self::Item, FancyError<Self::Error>>

Xor is just addition, with the requirement that x and y are mod 2.

fn negate(
&self,
ix: Option<SyncIndex>,
x: &Self::Item
) -> Result<Self::Item, FancyError<Self::Error>>

Negate by xoring x with 1.

fn and(
&self,
ix: Option<SyncIndex>,
x: &Self::Item,
y: &Self::Item
) -> Result<Self::Item, FancyError<Self::Error>>

And is just multiplication, with the requirement that x and y are mod 2.

fn or(
&self,
ix: Option<SyncIndex>,
x: &Self::Item,
y: &Self::Item
) -> Result<Self::Item, FancyError<Self::Error>>

Or uses Demorgan's Rule implemented with multiplication and negation.

fn and_many(
&self,
ix: Option<SyncIndex>,
args: &[Self::Item]
) -> Result<Self::Item, FancyError<Self::Error>>

Returns 1 if all wires equal 1.

fn or_many(
&self,
ix: Option<SyncIndex>,
args: &[Self::Item]
) -> Result<Self::Item, FancyError<Self::Error>>

Returns 1 if any wire equals 1.

fn mod_change(
&self,
ix: Option<SyncIndex>,
x: &Self::Item,
to_modulus: u16
) -> Result<Self::Item, FancyError<Self::Error>>

Change the modulus of x to to_modulus using a projection gate.

fn adder(
&self,
ix: Option<SyncIndex>,
x: &Self::Item,
y: &Self::Item,
carry_in: Option<&Self::Item>
) -> Result<(Self::Item, Self::Item), FancyError<Self::Error>>

Binary adder. Returns the result and the carry.

fn mux(
&self,
ix: Option<SyncIndex>,
b: &Self::Item,
x: &Self::Item,
y: &Self::Item
) -> Result<Self::Item, FancyError<Self::Error>>

If b=0 returns x else y.

fn mux_constant_bits(
&self,
ix: Option<SyncIndex>,
x: &Self::Item,
b1: bool,
b2: bool
) -> Result<Self::Item, FancyError<Self::Error>>

If x=0 return the constant b1 else return b2. Folds constants if possible.

fn outputs(
&self,
ix: Option<SyncIndex>,
xs: &[Self::Item]
) -> Result<(), FancyError<Self::Error>>

Output a slice of wires.